Move applications models to applications.py
